+++ src/aui-timepicker/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-* [AUI-1876](https://issues.liferay.com/browse/AUI-1876) AUI Timepicker selection should drop down to nearest value of current time
+* [AUI-1876](https://issues.liferay.com/browse/AUI-1876) AUI Timepicker selection should drop down to time nearest the time shown in its input field or the current local time if no time value is given
 
 ## [3.0.1](https://github.com/liferay/alloy-ui/releases/tag/3.0.1)